refactor fastify server
